Altera precisão do número informado

+++ Program.cs
@@ -1,4 +1,4 @@
-﻿double numero = 0;
+﻿int numero = 0;
 bool ehPar = false;
 
 Console.WriteLine("=============================");
@@ -7,7 +7,7 @@ Console.WriteLine("=============================");
 
 Console.Write("Informe um número:\t");
 
-numero = Convert.ToDouble(Console.ReadLine()!);
+numero = Convert.ToInt32(Console.ReadLine()!);
 ehPar = (numero % 2 == 0);
 
 if (ehPar)
